Cherry-pick: Find the current fullscreen video element rather than the one on top of stack
BUG=17007901

--------------------------------------------------------

Find the current fullscreen video element rather than the one on top of stack

When requestFullscren() is called, we don't know when willEnterFullscreen() will get called.

For android webview, requestFullscreen() might even result in no calls
to willenterfullscreen() if webapp doesn't support fullscreen video.
So before willEnterfullscreen is called, we should not use the provisional
video element on top of the stack to modify the layer tree.

Instead, we should wait for the provisional element to become current.
Therefore, we should call currentFullScreenElementFrom() on the fullscreen document.

BUG=389496

Review URL: https://codereview.chromium.org/399703002

Change-Id: I1817f26ae4113aeb8015046b321ebec567a98534
git-svn-id: svn://svn.chromium.org/blink/trunk@178714 bbb929c8-8fbe-4397-9dbb-9b2b20218538