Cherry-pick: Webkit setting for embedders that do not support fullscreen.
NOTE: this cherry-pick will only be landed after
https://codereview.chromium.org/290643005/

The new setting is fullscreenEnabled, but in this change
we have also added tests for
disallowFullscreenForNonMediaElements.

The motivation for this change is to support Android
WebView apps that do not implement WebChromeClient.onShowCustomView.
onShowCustomView is required to support fullscreen. When not supported,
we need to stop going fullscreen at the beginning of
FullscreenElementStack::requestFullScreenForElement
to avoid side effects and changes to the layout.

This failure to enter fullscreen falls into the following
condition on the living standard:
"There is a previously-established user preference,
security risk, or platform limitation."
http://fullscreen.spec.whatwg.org/

BUG=389496

Review URL: https://codereview.chromium.org/428633004

git-svn-id: svn://svn.chromium.org/blink/trunk@179470 bbb929c8-8fbe-4397-9dbb-9b2b20218538

Conflicts:
	Source/web/WebSettingsImpl.h
	public/web/WebSettings.h

Change-Id: I89a3dee15617b5858f4e727c88b0c5cfd655d18f